Added Window and Renderer classes to the codebase (#10)
* separate renderer

* Implemented classes for window creation and rendering functionality.

* Implemented the addition of 'width,' 'height,' and 'flags' getters for the Window class, enhancing its functionality. Set the renderer and window constructors as private to encapsulate the implementation details, and introduced 'createRenderer' and 'createWindow' functions for external instantiation. Accompanied the changes with comprehensive tests for the window. Additionally, addressed issues in the Engine test function to enhance overall code quality.

* change arrangement of arguments in createWindow function and remove default constructors

* try avoid cmake_build.yml actions

* changed region and window ptr type and added check for renderer resolve

* replace  unique_ptr to shared_ptr in renderer.hpp

* restore cmake_build.yml

* fixed window_test.cpp

+++ src/engine/src/sprite.cpp
@@ -11,8 +11,8 @@ Sprite::Sprite(Texture texture, Vec2 frameSize)
     int width{};
     int height{};
     SDL_QueryTexture(_texture.get(), nullptr, nullptr, &width, &height);
-    _rowCount = height / _frameSize[1];
-    _colCount = width / _frameSize[0];
+    _rowCount = height / static_cast<int>(_frameSize[1]);
+    _colCount = width / static_cast<int>(_frameSize[0]);
 }
 
 // TODO : renderer should be a member of Sprite
